Update versioning to 0.7.6
New with this version:
NominalApproach.get_param_scens now lets you get the scenarios in an approach associated with a varied parameter range. This essentially enables us to group some scenarios together (in other methods) to be treated as factors or as noise
plot.resilience_factor_comparison and tabulate.resilience_factor_comparison can now be used to compare the resilience of the system to different types of fault scenarios over a number of different defined factors
Replicates and randomization seeds can now be added in NominalApproach.add_param_ranges in case the parameter generation function has some stochasticity.